Update files for running liveblog with docker on windows
Restructured Dockerfile and replaced npm install with npm-install-que to enable the build on machines with slow internet and limited main memory.

Updated common.yml and removed the volume for mongoDB, because this is not working when docker is running in a VM.

Updated docker-compose.yml to include the mongoDB volume again.

Updated docker-compose-dev.yml to use a different command for startup.