Make the publish of embed an update instead of remove and add new file
Even when `boto3` library handles retries already,  I think there is a small margin of error here like how it
happen not long ago with missing embed files in clients websites. So the aim here is to reduce this margin.
The new approach is to try to replace the existing embed instead of removing the previous so, if there is a
network interruption the embed file would not end up missing.

LBSD-2535