Use proper edit-form.html view from superdesk-client-core
It seems a more recent version of this template was used to override superdesk UI and messed up
with user's profile pic not displaying it and also not allowing to change the picture.

I'm setting the proper view in place and in a upcoming commit will make the changes requested in the
ticket to make clear what we are overriding.

LBSD-1357