Archive blogs that are not able to delete because of syndication limitations
Blog that are syndicated and have active consumer should not be removed (at least not for now) given that other
instances might be trying to fetch data from them. So, in the case of blog marked for deletion, we are archiving them
instead to avoid getting error and being in the Deleted tab forever.

LBSD-2549