Maximize browser window on consumers_spec
Tests have started failing with an error that seems to be connected with the clickeable element not being visible in the viewport. The error is something like: Element is not clickable at point (411, 675). Other element would receive the click
To fix it I'm trying to maximize the browser window to make the element visible