Add checkbox to show/hide instagram embed description
Instagram embeds offer original posts caption. It also provides the opportunity to show or hide that caption. We are adding an option in liveblog editor to hide or show this description in the frontend side. Here the first step has been made by adding a checkbox to allow that actions to users

LBSD-2521