Hide users information if env var enabled
Given that we're creating a role with more powers for demo purposes (super editor), we need to care about the privacy of others users and do not expose data that can be sensitive. For this, we are implementing a new environment variable that if it's enabled, then the fields containing sensitive information will be replaced with a dummy string. This env var is False by default.

LBSD-2674