Refactor `publish_embed_to_s3` task
There were a couple of problems that have been fixed in this refactoring. Sometimes the `_publish_blog_embed_on_s3` function was provided without a theme (`None` value) which led to an error while generating the embed. While working on this refactoring I figured out that the theme parameter was really unnecessary therefore it has been removed. Now, to generate the embed, the function would use either the blog's theme or the output channel theme if provided.

The other issue that has been solved is that the `public_url` value of the blog was not being saved after the initial embed generation was processed. If the initial generation failed, no `public_url` was saved and subsequently it was not updated given a wrong logic. Now when the embed is generated and there is no `public_url` assigned, the blog will be updated anytime a new url is generated as part of the embed processing.

LBSD-2733