Implement use of features service in freetypes manager
The main menu configuration is run through `superdeskProvider` which is fully sync given that it is ready before anything else in the app (angular.config works like that). We cannot hide the freetypes manager entry from the menu based in the features service as the settings are loaded asynchronously. Because of that, I've needed to change the logic a little bit. Instead of hiding the menu, now we show it but explain the user that they need to upgrade their plan in order to have this feature in its full potential.

LBSD-2785