Change `newest_first` and `oldest_first` ordering attribute
There are some cases when the users create a post but do not publish it right away (like scheduled posts). Such posts, could end up not being rendered in the right position when using Newest First or Oldest First tabs in the readers side. To avoid such issues, we need to use `published_date` attribute instead as this is the one that stores the real date and time when the post is available to the readers

LBSD-2824